Ntm space compat (#4)
* 1.3.142 - spec fork compat

+++ src/main/java/enviromine/utils/EnviroUtils.java
@@ -6,7 +6,6 @@ import java.lang.reflect.Modifier;
 import java.math.BigDecimal;
 import java.util.ArrayList;
 
-import cpw.mods.fml.common.Loader;
 import net.minecraft.block.*;
 import org.apache.logging.log4j.Level;
 
@@ -29,11 +28,13 @@ import net.minecraftforge.common.IExtendedEntityProperties;
 import net.minecraftforge.common.util.ForgeDirection;
 import thaumcraft.common.blocks.BlockMagicalLeaves;
 
-import static enviromine.trackers.EnviroDataTracker.isTCLoaded;
+import static enviromine.core.EnviroMine.isTCLoaded;
 
 public class EnviroUtils
 {
+    //Get rid of unused methods
     public static final String[] reservedNames = new String[] {"CON", "COM", "PRN", "AUX", "CLOCK$", "NUL", "COM1", "COM2", "COM3", "COM4", "COM5", "COM6", "COM7", "COM8", "COM9", "LPT1", "LPT2", "LPT3", "LPT4", "LPT5", "LPT6", "LPT7", "LPT8", "LPT9"};
+                               //  ↑ scary
     public static final char[] specialCharacters = new char[] {'/', '\n', '\r', '\t', '\u0000', '\f', '`', '?', '*', '\\', '<', '>', '|', '\"', ':'};
 	private static final String IEEP_PLAYER_WITCHERY = "WitcheryExtendedPlayer";
 	private static final String IEEP_PLAYER_WITCHERY_CREATURE_TYPE = "CreatureType";
@@ -180,15 +181,15 @@ public class EnviroUtils
 	}
 
 
-	public static double getBiomeTemp(BiomeGenBase biome)
+	public static float getBiomeTemp(BiomeGenBase biome)
 	{
 		return getBiomeTemp(biome.temperature);
 	}
-	public static double getBiomeTemp(int x, int y, int z, BiomeGenBase biome)
+	public static float getBiomeTemp(int x, int y, int z, BiomeGenBase biome)
 	{
 		return getBiomeTemp(biome.getFloatTemperature(x, y, z));
 	}
-	private static double getBiomeTemp(float biomeTemp)
+	private static float getBiomeTemp(float biomeTemp)
 	{
 		// You can calibrate temperatures using these
 		// This does not take into account the time of day (These are the midday maximums)
@@ -196,7 +197,7 @@ public class EnviroUtils
 		float minTemp = -15F;
 
 		// CALCULATE!
-		return biomeTemp >= 0? Math.sin(Math.toRadians(biomeTemp*45F))*maxTemp : Math.sin(Math.toRadians(biomeTemp*45F))*minTemp;
+        return (float)(biomeTemp >= 0? Math.sin(Math.toRadians(biomeTemp*45F)) *maxTemp : Math.sin(Math.toRadians(biomeTemp*45F)) *minTemp) ;
 	}
 
 	/*
